pyproject.toml
[tool.poetry]
name = "autogen_docsearch"
version = "0.1.0"
description = "Docsearch with LLamaIndex and Autogen + Mermaid."
authors = ["Sean Dearnaley <SeanDearnaley@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
folder-tree-generator = "^0.1.13"
python-code-outline = "^0.1.7"
fastapi = "^0.101.0"
requests = "^2.31.0"
uvicorn = "^0.23.2"
tiktoken = "^0.4.0"
openai = "^0.27.8"
pyrate-limiter = "^2.10.0"
anthropic = "^0.3.9"
openai-functools = "^1.0.97"
rich = "^13.5.2"
loguru = "^0.7.0"
llama-index = "^0.8.41"
python-dotenv = "^1.0.0"
llama-hub = "^0.0.36"
sentencepiece = "^0.1.99"
torch = "^2.1.0"
transformers = "^4.34.0"
pillow = "^10.0.1"
pygithub = "^2.1.1"
docker = "^6.1.3"
pyautogen = "^0.1.7"
tenacity = "^8.2.3"

[tool.poetry.group.dev.dependencies]
pylint = "^2.17.5"
mypy = "^1.4.1"
pytest = "^7.4.0"
ruff = "^0.0.280"
black = "^23.7.0"
pytest-mock = "^3.10.0"
pytest-cov = "^4.1.0"
types-requests = "^2.31.0.1"
pylist = "^1.4.0"
protobuf = "^4.24.4"
flake8 = "^6.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
strict = true
files = "app"